request: Factor config rendering
